Chatbot/requirements.txt
langchain-openai
langchain-core
langchain-community
langchain-google-genai
streamlit
python-dotenv
